+++ src/saragurosnet/bussiness/__init__.py
@@ -1,12 +1,12 @@
 from src.chatbot.decisions_tree import DecisionsTree
 from src.chatbot.utils import format_fullname
 from src.common.cases import ConversationUseCases, MessageUseCases
-from src.common.logger import Logger
 from src.common.services import TwilioService
 from src.config import Config
 from src.saragurosnet.bussiness.context import Context
 from src.saragurosnet.bussiness.errors import say_error
 from src.saragurosnet.bussiness.initial_actions import group as initial_group
+from src.saragurosnet.bussiness.no_client_actions import group as no_client_group
 from src.saragurosnet.bussiness.preactions import group as pre_group
 from src.saragurosnet.types import MessageType, OptionType
 
@@ -23,67 +23,9 @@ tree.register_action_group(pre_group)
 tree.register_action_group(initial_group)
 # === End initial actions ===
 
-
-@tree.add_action("1.0", condition=lambda ctx: ctx.last_state == "0.2" and ctx.client != None and ctx.client.saraguros_id == None, next=["1.1", "1.2", "1.3"])
-def say_welcome_unknown(ctx: Context, id_func: str):
-    if ctx.client is None:
-        return say_error(ctx)
-
-    MessageUseCases().send_message(MessageType.WELCOME_UNKNOW.format(
-        name="Cliente"), ctx.event_twilio.from_number, ctx.conversation.id)
-    # twilio.send_message(MessageType.WELCOME_UNKNOW.format(name="Cliente"), receiver="whatsapp:" + ctx.client.phone)
-
-    ctx.last_state = id_func
-
-
-@tree.add_action("1.1", condition=lambda ctx: ctx.last_state == "1.0" and ctx.client != None and ctx.client.saraguros_id == None and (ctx.event_twilio.button_text == OptionType.PROMOTIONS or ctx.event_twilio.body == OptionType.PROMOTIONS), end=False, next="3.0")
-def send_promotions(ctx: Context, id_func: str):
-    if ctx.client is None:
-        return say_error(ctx)
-
-    MessageUseCases().send_message(MessageType.PROMOTIONS, ctx.event_twilio.from_number,
-                                   ctx.conversation.id, media_url=MediaUrlType.PROMOTIONS)
-
-    # twilio.send_message(MessageType.PROMOTIONS,
-    #                     receiver="whatsapp:" + ctx.client.phone,
-    #                     media_url=MediaUrlType.PROMOTIONS)
-
-    # fullname = format_fullname(ctx.client.names, ctx.client.lastnames)
-    # say_goodbye(fullname, "whatsapp:" + ctx.client.phone)
-    ctx.last_state = id_func
-
-
-@tree.add_action("1.2", condition=lambda ctx: ctx.last_state == "1.0" and ctx.client != None and ctx.client.saraguros_id == None and (ctx.event_twilio.button_text == OptionType.COVERAGES or ctx.event_twilio.body == OptionType.COVERAGES), end=False, next="3.0")
-def send_coverages(ctx: Context, id_func: str):
-    if ctx.client is None:
-        return say_error(ctx)
-
-    MessageUseCases().send_message(MessageType.COVERAGES, ctx.event_twilio.from_number,
-                                   ctx.conversation.id, media_url=MediaUrlType.COVERAGES)
-
-    # twilio.send_message(MessageType.COVERAGES,
-    #                     receiver="whatsapp:" + ctx.client.phone,
-    #                     media_url=MediaUrlType.COVERAGES)
-
-    # fullname = format_fullname(ctx.client.names, ctx.client.lastnames)
-    # say_goodbye(fullname, "whatsapp:" + ctx.client.phone)
-    ctx.last_state = id_func
-
-
-@tree.add_action("1.3", condition=lambda ctx: ctx.last_state == "1.0" and ctx.client != None and ctx.client.saraguros_id == None and (ctx.event_twilio.button_text == OptionType.AGENT or ctx.event_twilio.body == OptionType.AGENT))
-def talk_with_agent(ctx: Context, id_func: str):
-    if ctx.client is None:
-        return say_error(ctx)
-
-    fullname = format_fullname(ctx.client.names, ctx.client.lastnames)
-
-    MessageUseCases().send_message(MessageType.CONNECT_AGENT.format(
-        name=fullname), ctx.event_twilio.from_number, ctx.conversation.id)
-    # twilio.send_message(MessageType.CONNECT_AGENT.format(name=fullname), receiver="whatsapp:" + ctx.client.phone)
-
-    # TODO: Generar ticket de atencion al cliente
-    # TODO: Crear nuevo usuario en microwisp para tener su ID
-    # TODO: Enviar notificacion al agente con el ID del usuario
+# === No client actions ===
+tree.register_action_group(no_client_group)
+# === End no client actions ===
 
 
 @tree.add_action("2.0", condition=lambda ctx: ctx.last_state == "0.2" and ctx.client != None and ctx.client.saraguros_id != None)
